some work in e-commerce

+++ app/Http/Livewire/Admin/Category/Index.php
@@ -1,8 +1,10 @@
 <?php
 
 namespace App\Http\Livewire;
+
 namespace App\Http\Livewire\Admin\Category;
 
+use App\Models\Brand;
 use App\Models\Category;
 
 use Livewire\Component;
@@ -24,16 +26,31 @@ class Index extends Component
 
     public function destroyCategory()
     {
-        $category = Category::findOrFail($this->category_id);
+        $category = Category::find($this->category_id);
 
-        // Delete the Image also if available
-        $image_path = 'uploads/category/' . $category->image; {
-            if (File::exists($image_path));
-            File::delete($image_path);
+        if ($category) {
+            // Delete the Image also if available
+            $image_path = 'uploads/category/' . $category->image;
+
+            if (File::exists($image_path)) {
+                File::delete($image_path);
+            }
+
+            // Delete related brands if available
+            if ($category->brands()->count() > 0) {
+                $category->brands()->delete();
+                session()->flash('message', 'Category with Brands Deleted Successfully');
+            }
+
+            // Delete the category itself
+            $category->delete();
+            session()->flash('message', 'Category Deleted Successfully');
+            $this->dispatchBrowserEvent('close-modal');
+        } else {
+            session()->flash('message', 'No Category Found');
         }
-        $category->delete();
-        session()->flash('message', 'Category Deleted Successfully');
-        $this->dispatchBrowserEvent('close-modal');
+
+        return redirect()->back();
     }
 
     public function render()
